+++ dataSchema/numberSchema.go
@@ -3,14 +3,14 @@ package dataSchema
 import "errors"
 
 type Number struct {
-	Minimum          int16  `json:"minimum,omitempty"`          // (optional) Specifies a minimum numeric value, representing an inclusive lower limit. Only applicable for associated number or integer types.
-	Maximum          int16  `json:"maximum,omitempty"`          // (optional) Specifies a maximum numeric value, representing an inclusive upper limit. Only applicable for associated number or integer types.
+	Minimum          int    `json:"minimum,omitempty"`          // (optional) Specifies a minimum numeric value, representing an inclusive lower limit. Only applicable for associated number or integer types.
+	Maximum          int    `json:"maximum,omitempty"`          // (optional) Specifies a maximum numeric value, representing an inclusive upper limit. Only applicable for associated number or integer types.
 	ExculsiveMinimum int16  `json:"exclusiveMinimum,omitempty"` // (optional) Specifies a minimum numeric value, representing an exclusive lower limit. Only applicable for associated number or integer types.
 	ExclusiveMaximum int16  `json:"exclusiveMaximum,omitempty"` // (optional) Specifies a maximum numeric value, representing an exclusive upper limit. Only applicable for associated number or integer types.
 	MultipleOf       uint16 `json:"multipleOf,omitempty"`       // (optional) Specifies the multipleOf value number. The value must strictly greater than 0. Only applicable for associated number or integer types.
 }
 
-func NewNumber(defaultValue float64, unit string, minimum int16, maximum int16) Data {
+func NewNumber(defaultValue float64, unit string, minimum int, maximum int) Data {
 	return Data{
 		Default: defaultValue,
 		Type:    "number",
